chore(acpx): move runtime test fixtures to test-utils
